requirements.txt
pandas==2.1.4
sqlalchemy==1.4.49
psycopg2==2.9.6
apache-airflow==2.8.1